Sever/requirements.txt
fastapi
uvicorn
tensorflow
numpy
pillow
opencv-python
mediapipe
